requirements.txt
streamlit
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-google-genai
faiss-cpu
pypdf
python-dotenv
tenacity
docx2txt
openpyxl
rapidocr-onnxruntime
pillow
pandas
